abstracted location tracking - TrackMeSendDataActivity now inherits from TrackMeBaseActivity
