+++ app/src/main/res/layout/track_me.xml
@@ -33,12 +33,6 @@
         android:layout_alignBottom="@id/button_track_me_to_menu"
         android:layout_toEndOf="@+id/button_track_me_to_menu" />
 
-
-
-
-
-
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -75,7 +69,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/text_track_me_pace"
-        android:text="Average Pace: "
+        android:text="Current Pace: "
         android:textSize="25dp"
         android:textColor="@color/gray"
         android:textAlignment="center"
